requirements.txt
crewai>=0.28.0
langchain>=0.3.23
streamlit>=1.43.0
beautifulsoup4>=4.12.2
requests>=2.31.0
pdfminer.six>=20221105
python-docx>=0.8.11
selenium>=4.11.2
webdriver-manager>=4.0.0
openai>=0.27.8
PyPDF2>=3.0.1
crewai[tools]
google-generativeai
transformers>=4.51.3
duckduckgo-search
langchain-community>=0.3.21
torch>=2.6.0